+++ Tests/RxSwift/mapToTests.swift
@@ -14,7 +14,7 @@ import RxTest
 
 class MapToTests: XCTestCase {
 
-    let numbers: [Int?] = [1, nil, Int?(3)]
+    private let numbers: [Int?] = [1, nil, Int?(3)]
     private var observer: TestableObserver<String>!
 
     override func setUp() {
@@ -46,5 +46,74 @@ class MapToTests: XCTestCase {
             .completed(0)
         ])
         XCTAssertEqual(observer.events, correctValues)
+    }
+}
+
+// MARK: - Single
+extension MapToTests {
+    func testSingleReplaceSuccess() {
+        // Given
+        let expectedValue = "candy"
+        let scheduler = TestScheduler(initialClock: 0)
+        // When
+        let result = scheduler.start {
+            Single.just(1).mapTo(expectedValue).asObservable()
+        }
+        // Then
+        XCTAssertEqual(result.events, [
+            .next(TestScheduler.Defaults.subscribed, expectedValue),
+            .completed(TestScheduler.Defaults.subscribed)
+        ])
+    }
+
+    func testSingleNoReplaceFailure() {
+        // Given
+        let scheduler = TestScheduler(initialClock: 0)
+        // When
+        let result = scheduler.start {
+            Single<Int>.error(testError).mapTo("candy").asObservable()
+        }
+        // Then
+        XCTAssertEqual(result.events, [.error(TestScheduler.Defaults.subscribed, testError)])
+    }
+}
+
+// MARK: - Maybe
+extension MapToTests {
+    func testMaybeReplaceSuccess() {
+        // Given
+        let expectedValue = "candy"
+        let scheduler = TestScheduler(initialClock: 0)
+        // When
+        let result = scheduler.start {
+            Maybe.just(1).mapTo(expectedValue).asObservable()
+        }
+        // Then
+        XCTAssertEqual(result.events, [
+            .next(TestScheduler.Defaults.subscribed, expectedValue),
+            .completed(TestScheduler.Defaults.subscribed)
+        ])
+    }
+
+    func testMaybeNoReplaceFailure() {
+        // Given
+        let scheduler = TestScheduler(initialClock: 0)
+        // When
+        let result = scheduler.start {
+            Maybe<Int>.error(testError).mapTo("candy").asObservable()
+        }
+        // Then
+        XCTAssertEqual(result.events, [.error(TestScheduler.Defaults.subscribed, testError)])
+    }
+
+    func testMaybeNoReplaceEmpty() {
+        // Given
+        let scheduler = TestScheduler(initialClock: 0)
+        // When
+        let result = scheduler.start {
+            Maybe<Int>.empty().mapTo("candy").asObservable()
+        }
+        // Then
+        XCTAssertEqual(result.events, [.completed(TestScheduler.Defaults.subscribed)])
     }
 }