+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog
 - added `merge(with:)` for `Observable`
 - removed `flatMapSync` operator
 - added `apply` for `Completable` and `Maybe` 
+- added `mapTo` for `Single` and `Maybe`
+- added SPM support
 
 5.0.0
 -----